chore: move parabolaFit to new file
